MISUV-8334:   * optimise imports;   * small refactoring;
