+++ app/models/financialDetails/TransactionItem.scala
@@ -17,6 +17,7 @@
 package models.financialDetails
 
 import models.incomeSourceDetails.TaxYear
+import play.api.Logger
 import services.DateServiceInterface
 
 trait TransactionItem {
@@ -34,4 +35,20 @@ trait TransactionItem {
   val isLatePaymentInterest: Boolean
 
   def isOverdue()(implicit dateService: DateServiceInterface): Boolean
+
+  def getChargeTypeKey(codedOutEnabled: Boolean = false, reviewAndReconcileEnabled: Boolean = false): String =
+    (transactionType, subTransactionType) match {
+      case (PaymentOnAccountOne, _) => "paymentOnAccount1.text"
+      case (PaymentOnAccountTwo, _) => "paymentOnAccount2.text"
+      case (MfaDebitCharge, _) => "hmrcAdjustment.text"
+      case (BalancingCharge, Some(Nics2)) if codedOutEnabled => "class2Nic.text"
+      case (BalancingCharge, Some(Accepted)) if codedOutEnabled => "codingOut.text"
+      case (BalancingCharge, Some(Cancelled)) if codedOutEnabled => "cancelledPayeSelfAssessment.text"
+      case (BalancingCharge, _) => "balancingCharge.text"
+      case (PaymentOnAccountOneReviewAndReconcile, _) if reviewAndReconcileEnabled => "reviewAndReconcilePoa1.text"
+      case (PaymentOnAccountTwoReviewAndReconcile, _) if reviewAndReconcileEnabled => "reviewAndReconcilePoa2.text"
+      case error =>
+        Logger("application").error(s"Missing or non-matching charge type: $error found")
+        "unknownCharge"
+    }
 }
